refactor(env): Extract `wiki` into separate file
